Add priceRange parameter to search query
