--- react/constants/searchSortOptions.js
@@ -1,30 +1,0 @@
-export default [
-  {
-    value: 'OrderByTopSaleDESC',
-    label: 'ordenation.sales',
-  },
-  {
-    value: 'OrderByReleaseDateDESC',
-    label: 'ordenation.release.date',
-  },
-  {
-    value: 'OrderByBestDiscountDESC',
-    label: 'ordenation.discount',
-  },
-  {
-    value: 'OrderByPriceDESC',
-    label: 'ordenation.price.descending',
-  },
-  {
-    value: 'OrderByPriceASC',
-    label: 'ordenation.price.ascending',
-  },
-  {
-    value: 'OrderByNameASC',
-    label: 'ordenation.name.ascending',
-  },
-  {
-    value: 'OrderByNameDESC',
-    label: 'ordenation.name.descending',
-  },
-]
